made changes in message_slot.c

+++ message_slot.h
@@ -2,7 +2,7 @@
 #define MESSAGE_SLOT
 
 #include <linux/ioctl.h>
-#include <stdlib.h>
+//#include <stdlib.h>
 
 #define MAJOR_NUM 235
 #define MSG_SLOT_CHANNEL (_IOW(MAJOR_NUM, 0, unsigned long)) 
@@ -15,7 +15,7 @@
 typedef struct channel_node
 {
     struct channel_node* next;
-    size_t msg_len; // actual length of message - bounded by 128 bytes
+    int msg_len; // actual length of message - bounded by 128 bytes
     char msg[BUF_LEN]; // the message. Note that the size is already set at instantiation (it is O(1))
     unsigned long id; // this not the message_slot minor, so is not bounded in size
 } ch_node ;
